+++ src/calculator/v2.js
@@ -1,0 +1,5 @@
+const { Base, ...passThrough } = require('./base');
+
+class Calculator extends Base {}
+
+module.exports = { Calculator, ...passThrough };