+++ src/index.js
@@ -1,0 +1,4 @@
+module.exports = {
+  Processor: require('./processor'),
+  Error: require('./error')
+};