+++ src/processor.js
@@ -1,39 +1,10 @@
 const debug = require('debug')('iiif-processor:main');
 const mime = require('mime-types');
+const path = require('path');
 const sharp = require('sharp');
-const { Formats, Qualities, Operations, pathToIiif } = require('./lib/transform');
-const IIIFError = require('./lib/error');
-const IIIFProfile = [
-  'http://iiif.io/api/image/2/level2.json',
-  {
-    formats: Formats,
-    qualities: Qualities,
-    supports: [
-      'regionByPx',
-      'sizeByW',
-      'sizeByWhListed',
-      'cors',
-      'regionSquare',
-      'sizeByDistortedWh',
-      'sizeAboveFull',
-      'canonicalLinkHeader',
-      'sizeByConfinedWh',
-      'sizeByPct',
-      'jsonldMediaType',
-      'regionByPct',
-      'rotationArbitrary',
-      'sizeByH',
-      'baseUriRedirect',
-      'rotationBy90s',
-      'profileLinkHeader',
-      'sizeByForcedWh',
-      'sizeByWh',
-      'mirroring'
-    ]
-  }
-];
-
-const DefaultPathPrefix = '/iiif/2/';
+const { Operations } = require('./transform');
+const IIIFError = require('./error');
+const IIIFVersions = require('./versions');
 
 const fixupSlashes = (path, leaveOne) => {
   const replacement = leaveOne ? '/' : '';
@@ -41,41 +12,34 @@ const fixupSlashes = (path, leaveOne) => {
 };
 
 class Processor {
-  constructor (url, streamResolver, ...args) {
-    const opts = this.parseOpts(args);
-
+  constructor (version, url, streamResolver, opts = {}) {
     if (typeof streamResolver !== 'function') {
       throw new IIIFError('streamResolver option must be specified');
     }
 
-    this
-      .setOpts(opts)
-      .initialize(url, streamResolver);
-  }
+    if (opts.max?.height && !opts.max?.width) {
+      throw new IIIFError('maxHeight cannot be specified without maxWidth');
+    };
 
-  parseOpts (args) {
-    if (args.length >= 1 && typeof args[0] === 'function') {
-      console.warn('Passing dimensionFunction, maxWidth, and includeMetadata as separate parameters is deprecated. Please see the documentation.');
-      return {
-        dimensionFunction: args.shift(),
-        maxWidth: args.shift(),
-        includeMetadata: args.shift()
-      };
-    } else if (args.length === 1 && typeof args[0] === 'object') {
-      return args[0];
-    } else {
-      return {};
-    }
+    const defaults = {
+      pathPrefix: `/iiif/${version}/`,
+      dimensionFunction: this.defaultDimensionFunction,
+      density: null
+    };
+
+    this
+      .setOpts({ ...defaults, ...opts })
+      .initialize(version, url, streamResolver);
   }
 
   setOpts (opts) {
     this.errorClass = IIIFError;
-    this.dimensionFunction = opts.dimensionFunction || this.defaultDimensionFunction;
-    this.maxWidth = opts.maxWidth;
+    this.dimensionFunction = opts.dimensionFunction;
+    this.max = { ...opts.max };
     this.includeMetadata = !!opts.includeMetadata;
-    this.density = opts.density || null;
-    this.pathPrefix = fixupSlashes(opts.pathPrefix, true) || DefaultPathPrefix;
-    this.sharpOptions = opts.sharpOptions || {};
+    this.density = opts.density;
+    this.pathPrefix = fixupSlashes(opts.pathPrefix, true);
+    this.sharpOptions = { ...opts.sharpOptions };
 
     return this;
   }
@@ -83,13 +47,16 @@ class Processor {
   parseUrl (url) {
     const parser = new RegExp(`(?<baseUrl>https?://[^/]+${this.pathPrefix})(?<path>.+)$`);
     const { baseUrl, path } = parser.exec(url).groups;
-    const result = pathToIiif(path);
+    const result = this.Implementation.Calculator.parsePath(path);
     result.baseUrl = baseUrl;
 
     return result;
   }
 
-  initialize (url, streamResolver) {
+  initialize (version, url, streamResolver) {
+    this.version = version;
+    this.Implementation = IIIFVersions[this.version];
+
     const params = this.parseUrl(url);
     debug('Parsed URL: %j', params);
     Object.assign(this, params);
@@ -157,40 +124,34 @@ class Processor {
       sizes.push({ width: size[0], height: size[1] });
     }
 
-    const doc = this.infoDoc(dim, sizes);
+    const id = [fixupSlashes(this.baseUrl), fixupSlashes(this.id)].join('/');
+    const doc = this.Implementation.infoDoc({ id, ...dim, sizes, max: this.max });
+    for (const prop in doc) {
+      if (doc[prop] === null || doc[prop] === undefined) delete doc[prop];
+    }
 
-    if (this.maxWidth) doc.profile[1].maxWidth = this.maxWidth;
-
-    return { contentType: 'application/json', body: JSON.stringify(doc) };
+    // Serialize sets as arrays
+    const body = JSON.stringify(doc, (_key, value) =>
+      value?.constructor === Set ? [...value] : value
+    );
+    return { contentType: 'application/json', body };
   }
 
-  infoDoc (dim, sizes) {
-    return {
-      '@context': 'http://iiif.io/api/image/2/context.json',
-      '@id': [fixupSlashes(this.baseUrl), fixupSlashes(this.id)].join('/'),
-      protocol: 'http://iiif.io/api/image',
-      width: dim.width,
-      height: dim.height,
-      sizes,
-      tiles: [{ width: 512, height: 512, scaleFactors: sizes.map((_v, i) => 2 ** i) }],
-      profile: IIIFProfile
-    };
-  }
-
-  async pipeline (dim) {
-    return new Operations(dim, this.sharpOptions)
+  operations (dim) {
+    const { sharpOptions: sharp, max } = this;
+    return new Operations(this.version, dim, { sharp, max })
       .region(this.region)
       .size(this.size)
       .rotation(this.rotation)
       .quality(this.quality)
       .format(this.format, this.density)
-      .withMetadata(this.includeMetadata)
-      .pipeline();
+      .withMetadata(this.includeMetadata);
   }
 
   async iiifImage () {
     const dim = await this.dimensions();
-    const pipeline = await this.pipeline(dim);
+    const operations = this.operations(dim);
+    const pipeline = await operations.pipeline();
 
     const result = await this.withStream({ id: this.id, baseUrl: this.baseUrl }, async (stream) => {
       debug('piping stream to pipeline');
@@ -199,7 +160,15 @@ class Processor {
       return await transformed.toBuffer();
     });
     debug('returning %d bytes', result.length);
-    return { contentType: mime.lookup(this.format), body: result };
+    debug('baseUrl', this.baseUrl);
+
+    const canonicalUrl = new URL(path.join(this.id, operations.canonicalPath()), this.baseUrl);
+    return {
+      canonicalLink: canonicalUrl.toString(),
+      profileLink: this.Implementation.profileLink,
+      contentType: mime.lookup(this.format),
+      body: result
+    };
   }
 
   async execute () {
@@ -211,4 +180,4 @@ class Processor {
   }
 }
 
-module.exports = { Processor, IIIFError };
+module.exports = Processor;