+++ tests/v2/processor.test.js
@@ -2,7 +2,8 @@
 
 const assert = require('assert');
 const { Stream } = require('stream');
-const iiif = require('../index');
+const IIIFError = require('../../src/error');
+const Processor = require('../../src/processor');
 
 let subject;
 const base = 'https://example.org/iiif/2/ab/cd/ef/gh/i';
@@ -10,7 +11,7 @@ const dims = [{ width: 1024, height: 768 }];
 
 describe('IIIF Processor', () => {
   beforeEach(() => {
-    subject = new iiif.Processor(`${base}/10,20,30,40/pct:50/45/default.png`, ({ id }) => id);
+    subject = new Processor(2, `${base}/10,20,30,40/pct:50/45/default.png`, ({ id }) => id);
   });
 
   it('Parse URL', () => {
@@ -24,7 +25,7 @@ describe('IIIF Processor', () => {
   });
 
   it('Create pipeline', async () => {
-    const pipe = await subject.pipeline(dims);
+    const pipe = await subject.operations(dims).pipeline();
     const opts = pipe.options;
 
     assert.strictEqual(opts.width, 15);
@@ -41,7 +42,7 @@ describe('IIIF Processor', () => {
 
 describe('Include metadata', () => {
   beforeEach(() => {
-    subject = new iiif.Processor(
+    subject = new Processor(2, 
       `${base}/10,20,30,40/pct:50/45/default.jpg`,
       ({ id }) => id,
       { includeMetadata: true }
@@ -49,7 +50,7 @@ describe('Include metadata', () => {
   });
 
   it('Includes preexisting metadata', async () => {
-    const pipe = await subject.pipeline(dims);
+    const pipe = await subject.operations(dims).pipeline();
     const opts = pipe.options;
 
     assert.strictEqual(opts.withMetadata, true);
@@ -58,11 +59,11 @@ describe('Include metadata', () => {
 
 describe('TIFF Download', () => {
   beforeEach(() => {
-    subject = new iiif.Processor(`${base}/10,20,30,40/pct:50/45/default.tif`, ({ id }) => id);
+    subject = new Processor(2, `${base}/10,20,30,40/pct:50/45/default.tif`, ({ id }) => id);
   });
 
   it('Output TIFF format', async () => {
-    const pipe = await subject.pipeline(dims);
+    const pipe = await subject.operations(dims).pipeline();
     const opts = pipe.options;
 
     assert.strictEqual(opts.width, 15);
@@ -79,7 +80,7 @@ describe('TIFF Download', () => {
 describe('Density', () => {
   beforeEach(() => {
     subject = (ext) => {
-      return new iiif.Processor(
+      return new Processor(2, 
         `https://example.org/iiif/2/ab/cd/ef/gh/i/10,20,30,40/pct:50/45/default.${ext}`,
         ({ id }) => id,
         { density: 600 }
@@ -89,7 +90,7 @@ describe('Density', () => {
 
   it('Adds density to TIFF', async () => {
     const processor = subject('tif');
-    const pipe = await processor.pipeline(dims);
+    const pipe = await processor.operations(dims).pipeline();
     const opts = pipe.options;
 
     assert.strictEqual(opts.tiffXres, 600 / 25.4);
@@ -98,7 +99,7 @@ describe('Density', () => {
 
   it('Adds density to JPEG', async () => {
     const processor = subject('jpg');
-    const pipe = await processor.pipeline(dims);
+    const pipe = await processor.operations(dims).pipeline();
     const opts = pipe.options;
 
     assert.strictEqual(opts.withMetadataDensity, 600);
@@ -106,7 +107,7 @@ describe('Density', () => {
 
   it('Adds density to PNG', async () => {
     const processor = subject('png');
-    const pipe = await processor.pipeline(dims);
+    const pipe = await processor.operations(dims).pipeline();
     const opts = pipe.options;
 
     assert.strictEqual(opts.withMetadataDensity, 600);
@@ -115,54 +116,36 @@ describe('Density', () => {
 
 describe('constructor', () => {
   it('must parse the object-based constructor', () => {
-    subject = new iiif.Processor(
+    subject = new Processor(2, 
       `${base}/10,20,30,40/pct:50/45/default.tif`,
       () => 'streamResolver',
-      { dimensionFunction: () => 'dimensionFunction', maxWidth: 'maxWidth', includeMetadata: true, density: 600 }
+      { dimensionFunction: () => 'dimensionFunction', max: { width: 'maxWidth' }, includeMetadata: true, density: 600 }
     );
 
     assert.strictEqual(subject.streamResolver(), 'streamResolver');
     assert.strictEqual(subject.dimensionFunction(), 'dimensionFunction');
-    assert.strictEqual(subject.maxWidth, 'maxWidth');
+    assert.strictEqual(subject.max.width, 'maxWidth');
     assert.strictEqual(subject.includeMetadata, true);
     assert.strictEqual(subject.density, 600);
-  });
-
-  it('must parse the multi-argument constructor', () => {
-    const spy = jest.spyOn(global.console, 'warn').mockImplementation();
-    subject = new iiif.Processor(
-      `${base}/10,20,30,40/pct:50/45/default.tif`,
-      () => 'streamResolver',
-      () => 'dimensionFunction',
-      'maxWidth',
-      true
-    );
-    expect(spy).toHaveBeenCalledWith(expect.stringMatching(/deprecated/));
-    spy.mockRestore();
-
-    assert.strictEqual(subject.streamResolver(), 'streamResolver');
-    assert.strictEqual(subject.dimensionFunction(), 'dimensionFunction');
-    assert.strictEqual(subject.maxWidth, 'maxWidth');
-    assert.strictEqual(subject.includeMetadata, true);
-    assert.strictEqual(subject.density, null);
   });
 
   it('properly handles custom sharp options', async () => {
     let pipe;
     
-    subject = new iiif.Processor(
+    subject = new Processor(2,
       `${base}/10,20,30,40/pct:50/45/default.tif`,
-      () => 'streamResolver'
+      () => 'streamResolver',
+      { sharpOptions: { sequentialRead: false } }
     );
-    pipe = await subject.pipeline(dims);
+    pipe = await subject.operations(dims).pipeline();
     assert.strictEqual(pipe.options.input.sequentialRead, false);
     
-    subject = new iiif.Processor(
+    subject = new Processor(2, 
       `${base}/10,20,30,40/pct:50/45/default.tif`,
       () => 'streamResolver',
       { sharpOptions: { sequentialRead: true } }
     );
-    pipe = await subject.pipeline(dims);
+    pipe = await subject.operations(dims).pipeline();
     assert.strictEqual(pipe.options.input.sequentialRead, true);
   })
 });
@@ -170,14 +153,14 @@ describe('constructor', () => {
 describe('constructor errors', () => {
   it('requires a streamResolver', () => {
     assert.throws(() => {
-      return new iiif.Processor(`${base}/10,20,30,40/pct:50/45/default.tif`, {});
-    }, iiif.IIIFError);
+      return new Processor(2, `${base}/10,20,30,40/pct:50/45/default.tif`, {});
+    }, IIIFError);
   });
 
   it('requires a valid URL', () => {
     assert.throws(() => {
-      return new iiif.Processor(`${base}/10,20,30,40/pct:50/45/default.blargh`, ({ id }) => id);
-    }, iiif.IIIFError);
+      return new Processor(2, `${base}/10,20,30,40/pct:50/45/default.blargh`, ({ id }) => id);
+    }, IIIFError);
   });
 });
 
@@ -194,7 +177,7 @@ describe('stream processor', () => {
       });
     }
 
-    const subject = new iiif.Processor(`https://example.org/iiif/2/ab/cd/ef/gh/i/10,20,30,40/pct:50/45/default.png`, streamResolver, {pathPrefix: 'iiif/2/ab/cd/ef/gh'});
+    const subject = new Processor(2, `https://example.org/iiif/2/ab/cd/ef/gh/i/10,20,30,40/pct:50/45/default.png`, streamResolver, {pathPrefix: 'iiif/2/ab/cd/ef/gh'});
     subject.execute();
   })
 })
@@ -215,7 +198,7 @@ describe('dimension function', () => {
       return { w: 100, h: 100 }
     }
 
-    const subject = new iiif.Processor(
+    const subject = new Processor(2, 
       `https://example.org/iiif/2/ab/cd/ef/gh/i/10,20,30,40/pct:50/45/default.png`,
       streamResolver,
       { dimensionFunction, pathPrefix: 'iiif/2/ab/cd/ef/gh' }